RUM-16718: Created a specific DefaultSwiftUIRUMActionsPredicate for macOS.
The Mac version does not need the `isLegacyDetectionEnabled` parameter.
