+++ DatadogRUM/Sources/Instrumentation/Actions/SwiftUI/iOS/SwiftUIRUMActionsPredicate_iOS.swift
@@ -1,0 +1,38 @@
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2019-Present Datadog, Inc.
+ */
+
+#if !os(macOS)
+import UIKit
+import DatadogInternal
+
+/// Default implementations of `SwiftUIRUMActionsPredicate`
+public struct DefaultSwiftUIRUMActionsPredicate {
+    /// Whether to enable SwiftUI action detection on iOS 17 and below.
+    /// When set to `false`, actions will only be detected on iOS 18+ where the detection is more reliable.
+    /// Defaults to `true` for backward compatibility.
+    private let isLegacyDetectionEnabled: Bool
+
+    /// Creates a default SwiftUI RUM actions predicate.
+    /// - Parameter isLegacyDetectionEnabled: Whether to enable SwiftUI action detection on iOS 17 and below.
+    ///     Set to `true` to enable legacy detection, which may capture more actions but with less reliability.
+    ///     Set to `false` to only use the more reliable iOS 18+ detection.
+    public init(isLegacyDetectionEnabled: Bool) {
+        self.isLegacyDetectionEnabled = isLegacyDetectionEnabled
+    }
+}
+
+// MARK: DefaultSwiftUIRUMActionsPredicate
+extension DefaultSwiftUIRUMActionsPredicate: SwiftUIRUMActionsPredicate {
+    public func rumAction(with componentName: String) -> RUMAction? {
+        if #available(iOS 18.0, *) {
+            return RUMAction(name: componentName, attributes: [:])
+        } else if isLegacyDetectionEnabled {
+            return RUMAction(name: componentName, attributes: [:])
+        }
+        return nil
+    }
+}
+#endif
