+++ DatadogRUM/Sources/Instrumentation/Actions/SwiftUI/macOS/SwiftUIRUMActionsPredicate_macOS.swift
@@ -1,0 +1,19 @@
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2019-Present Datadog, Inc.
+ */
+
+#if os(macOS)
+import AppKit
+import DatadogInternal
+
+/// Default implementations of `SwiftUIRUMActionsPredicate`
+public struct DefaultSwiftUIRUMActionsPredicate: SwiftUIRUMActionsPredicate {
+    public init() {}
+
+    public func rumAction(with componentName: String) -> RUMAction? {
+        return RUMAction(name: componentName, attributes: [:])
+    }
+}
+#endif
