+++ DatadogRUM/Sources/Instrumentation/Views/SwiftUI/SwiftUIRUMViewsPredicate+objc.swift
@@ -63,11 +63,19 @@ public protocol objc_SwiftUIRUMActionsPredicate: AnyObject {
 public class objc_DefaultSwiftUIRUMActionsPredicate: NSObject, objc_SwiftUIRUMActionsPredicate {
     private let swiftPredicate: DefaultSwiftUIRUMActionsPredicate
 
+    #if os(macOS)
+    @objc(init)
+    override public init() {
+        swiftPredicate = DefaultSwiftUIRUMActionsPredicate()
+        super.init()
+    }
+    #else
     @objc(initWithIsLegacyDetectionEnabled:)
     public init(isLegacyDetectionEnabled: Bool) {
         swiftPredicate = DefaultSwiftUIRUMActionsPredicate(isLegacyDetectionEnabled: isLegacyDetectionEnabled)
         super.init()
     }
+    #endif
 
     public func rumAction(with componentName: String) -> objc_RUMAction? {
         swiftPredicate.rumAction(with: componentName).map {
